feat: ✨ Moved Stripe functions to extension
